services/master/pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bedantH", email = "bedanthota@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "docker>=7.1.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=5.0.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "shared",
    "sqlmodel>=0.0.31",
    "watchfiles>=1.0.0",
]

[project.scripts]
master = "master:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
shared = { workspace = true }
